requirements.txt
absl-py==2.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astroid==4.0.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
cachetools==6.2.1
cattrs==25.3.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
db-dtypes==1.4.3
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.120.1
fastjsonschema==2.21.2
flatbuffers==25.2.10
fonttools==4.60.1
fqdn==1.5.1
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.8.2
google-auth==2.41.1
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.18.0
google-cloud-core==2.4.3
google-cloud-storage==2.14.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.56.4
grpcio==1.76.0
h11==0.16.0
h5py==3.11.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipdb==0.13.13
ipykernel==7.0.1
ipython==8.37.0
isoduration==20.11.0
isort==7.0.0
jedi==0.19.2
jh2==5.0.10
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server-mathjax==0.2.6
jupyter-server-ydoc==2.1.1
jupyter-ydoc==3.1.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_fileid==0.9.3
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_code_formatter==3.0.2
jupyterlab_execute_time==3.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
lark==1.3.0
libclang==18.1.1
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbdime==4.0.2
nbformat==5.10.4
nest-asyncio==1.6.0
niquests==3.15.2
notebook==7.4.5
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
openmeteo_requests==1.7.3
openmeteo_sdk==1.21.2
opt_einsum==3.4.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==3.20.3
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycrdt==0.12.42
pycrdt-store==0.1.2
pycrdt-websocket==0.16.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pylint==4.0.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qh3==1.5.5
referencing==0.37.0
requests==2.32.5
requests-cache==1.2.1
requests-oauthlib==2.0.0
retry-requests==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.28.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlite-anyio==0.2.3
stack-data==0.6.3
starlette==0.49.0
tensorboard==2.19.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.3.0
tomlkit==0.13.3
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
url-normalize==2.2.1
urllib3==2.5.0
urllib3-future==2.14.905
uvicorn==0.28.1
wassima==2.0.2
wcwidth==0.2.14
wear_today==0.0.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
wrapt==2.0.0
timezonefinder==8.1.0
geopy==2.4.1
